Rain changes #1.5 length + speed
The last change to the file was missing a line of code that would allow the program to randomise the size of the raindrops between 5 and 15 pixels.

float len = random(5, 15);
//in line 5//

&

line(x,y,x,y+len);
//in line 17//

Next comes the edits to the speed. the speed was randomised in order to create a more life like apearance to the animation.

float yspeed = random(4, 5); //on line 3//

I also increased the position that drops would begin animating outside of the parameters of the screen, so that by the time they came into view they were falling in a less uniform manner

float y = random(-600, 200);
//line 4//