Randomisation of speed
At this stage the rain is pretty much all but ready to go, however I felt like it could be improved and there were a lot of helpful tutorials online that helped me do just that.

My main issues with the rain at this point are three things

1. The Rain falls a little bit too uniformily

2. The speed is too organised

3. The droplets all look the same

I learnt about the map() function, which allows you store a range of values and tells the program to choose and implement code within the parameters of the range of numbers listed in map e.g. map(1, 20); //but with more values

In order to create a randomised speed for the droplets I created a new float called 'z' which would choose a random value between 1 and 20.

The value was then fed into 'yspeed' via a map function, allowing the code to choose a random speed at which the drops would fall, whilst the rest of the values would represent the distance between drops & the amount falling at the same time.