Reversion to previous rain and adjusted rain speed
After messing around a bit with the code  specifically the thickness and the speed, i decided that my initial code produced a more appealing result and was closer to the appearance of real rain which is what I was hoping to do to begin with.

I edited the thicknesses using

void fall() {
float thick = map(z, 0, 15, 1, 3);
    strokeWeight(thick);
}

Using the previously established float 'z', producing a random number between 0, 20,  this in conjunction with 'z' would create a piece of code that creates rain drops of random thickness between the range of 1 to 3 pixels.

Instead I simply adjusted the previous code using random(0, 20) as input for the rain drop speed and i decided to keep the rain drops at there standard thickness so as not to obstruct the background image i'm hoping to put in at the end