Mouse functions intergrated!
At this stage i'm uploading all the up-to-date current code in the latest commit, at the same time. The clock is ready now the mouse clicks now register the change from date to month without the neccesity of holding the mouse down.

I may need to add an extra section in order to display the weather as a seperate screen because it slightly obstructive when your trying to look at the date.