Weather Image: Clock Finished!
At this stage the clock is pretty much completed.

The Time, Date & Weather are in the correct order.

The mouse click registers the transition to the next set of data.

The rain falling resembles real rain and the image compliments it nicely

And the final mouse click returns the screen to the first function The Date.

In order to create an image I went through a bit of trial and error (much like most stages of this project) thanks to a helpful internet tutorial that introduced me the PImage datatype which allows people to store and display images within their processing program.

PImage img;
//used to declare the image datatype

img = loadImage("86966_cloud.png");
//after which I added this line of code into the setup function nesting it inside the for loop housing the code that displayed the rain, which means everytime the rain is shown so is the image

image(img, width / 2 - 63, 0);
//this line was added into the 'drawDots' function allowing to insure the image would be called whenever the rain was present.
After which I used a bit f trial and error to figure out the correct pixel co-oridnates to place the cloud image in the centre of the display.

 if (numClicks == 0) {
    drawTime(); }
  else if (numClicks == 1) {
    drawDate();
    }
   else if (numClicks == 2) {
     drawDots();
   }
//this is the section I used to tell the program how to register and use mouse clicks after which they would be implemented using this code

void mousePressed() {
  numClicks++;
  if(numClicks > 2) numClicks = 0;

If i think of any more features I will return and see about updating this code